pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lrPymRPC"
version = "0.2.5"
description = "gRPC client of Python module for LogicResearch-Tools."
authors = [{ name = "LogicResearch K.K.(MATSUDA Masahiro)", email = "matsuda@logic-research.co.jp" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "grpcio",
  "grpcio-tools"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lrPymRPC","lrPymRPC.script","lrPymRPC.proto"]

[project.scripts]
lrRPC-client = "lrPymRPC.client:main"
